requirements.txt
Django==4.1.3
djangorestframework==3.14.0
numpy
pandas
  scikit-learn
pip install joblib
